Reorganize existing hooks
